Add examples:seed Rake task.
I got tired of trying to do a make-shift system like this, so I threw
together a little Rake task to do it for me. This task will clone some
git repos for gems that have YARD docs and then run sord and generate
the rbi files for each.

This also includes an examples:reset Rake task to delete the
sord_examples directory.