Update seeding tasks in Rakefile.
Add oga to list of repositories cloned as part of the seeding.

Run yard with `bundle exec` for consistency and with `system` so it
outputs information to the log. This helps with debugging if something
breaks.

Run commands inside a clean Bundler env so `bundle exec` doesn't cause problems.

Add some colors to messages output by the Rake tasks.