Fix the type signature with circular inheritance
If you define a class with the same name as the parent class, sord will
generate the type signature with circular inheritance.
For example, the code is as follows.

```ruby
class ApplicationRecord
end

module Api
  class ApplicationRecord < ::ApplicationRecord
  end
end
```

This commit adds a name separator when needed and fixes the type signature
so that it is not circular.